+++ CHANGELOG.md
@@ -29,6 +29,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Add `MissedPredicatesPerTSCMetric`.
+- Add `identifier` parameter to `TSCDsl`.
+
 
 ## [1.0] - 14.10.2025